// src/frontend-applications/agent-portal/model/configuration/TableHeaderHeight.ts
/**
 * Copyright (C) 2006-2024 KIX Service Software GmbH, https://www.kixdesk.com
 * --
 * This software comes with ABSOLUTELY NO WARRANTY. For details, see
 * the enclosed file LICENSE for license information (GPL3). If you
 * did not receive this file, see https://www.gnu.org/licenses/gpl-3.0.txt.
 * --
 */

export enum TableHeaderHeight {

    SMALL = 1.75,

    LARGE = 2.25

}